Bigpoint Gasstation programming exercise
The goal is to implement the interface of a gas station which has
multiple gas pumps. Each

gas pump provides a specific amount of gas of a certain gas type.

Additionally, customers are requesting gas with different requirements
regarding the

amount, type of gas and price they are willing to pay for it. Your
implementation has to

take all this into account and serve gas when available. For further
information please refer to the

documentation of our classes.